Fixed some integer casts

+++ WFInfo/SnapItOverlay.xaml.cs
@@ -80,8 +80,8 @@ namespace WFInfo
             }
 
             Bitmap cutout = tempImage.Clone(new Rectangle((int)(topLeft.X * OCR.dpiScaling), (int)(topLeft.Y * OCR.dpiScaling), (int)(rectangle.Width * OCR.dpiScaling), (int)(rectangle.Height * OCR.dpiScaling)), System.Drawing.Imaging.PixelFormat.DontCare);
-            int xPos = topLeft.X + (int)rectangle.Width / 2 * (int)(OCR.dpiScaling);
-            int yPos = topLeft.Y + 10 * (int)(OCR.dpiScaling);
+            int xPos = (int)((topLeft.X + rectangle.Width / 2) * OCR.dpiScaling);
+            int yPos = (int)((topLeft.Y + 10) * OCR.dpiScaling);
             Task.Factory.StartNew(() => OCR.ProcessSnapIt(cutout,tempImage,xPos,yPos));
 
             closeOverlay();
